requirements.txt
django
django-extensions
faker
djangorestframework